# backend2/docs/openapi.yaml
openapi: 3.0.3
info:
  title: e-KisanSetu Farmer Procurement & Live Mandi Queue API
  version: 1.0.0
  description: |
    Comprehensive REST & Real-Time Server-Sent Events (SSE) API specification for the National Smart Farmer Procurement Management System.
    Powers APMC procurement centers, token issuance, dynamic FIFO queue ordering, multi-stage procurement tracking, ML waiting-time predictions, and multi-channel notifications.
servers:
  - url: http://localhost:5000/api/v1
    description: Local Development Server
  - url: http://localhost:5000
    description: Root Server (Direct Route Aliases)

paths:
  /centers:
    get:
      summary: List all procurement centers
      description: Retrieve all APMC Mandis and Procurement Centers with optional filters for state, district, crop, and status.
      parameters:
        - name: state
          in: query
          schema: { type: string }
          example: Uttar Pradesh
        - name: district
          in: query
          schema: { type: string }
          example: Ghaziabad
        - name: status
          in: query
          schema: { type: string, enum: [open, busy, closed] }
          example: open
        - name: crop
          in: query
          schema: { type: string }
          example: Wheat
      responses:
        '200':
          description: List of centers retrieved successfully

  /centers/{id}/schedule:
    get:
      summary: Get center operating schedule
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          example: up_c1
        - name: startDate
          in: query
          schema: { type: string, format: date }
          example: "2026-09-05"
        - name: days
          in: query
          schema: { type: integer, default: 7 }
      responses:
        '200':
          description: Operating schedule returned
    put:
      summary: Update center operating schedule (Admin only)
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          example: up_c1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                scheduleDate: { type: string, format: date, example: "2026-09-05" }
                openTime: { type: string, example: "07:30" }
                closeTime: { type: string, example: "18:30" }
                totalSlots: { type: integer, example: 90 }
                isHoliday: { type: boolean, example: false }
                notes: { type: string, example: "Extended harvest procurement shift" }
      responses:
        '200':
          description: Schedule updated successfully

  /centers/{id}/queue:
    get:
      summary: Get live queue at procurement center
      description: Returns all farmers currently in procurement, called, waiting in order, and recently completed today.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          example: up_c1
      responses:
        '200':
          description: Live center queue details
    put:
      summary: Administrative queue reordering (Admin only)
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reorderList]
              properties:
                reorderList:
                  type: array
                  items:
                    type: object
                    required: [tokenId, newPosition]
                    properties:
                      tokenId: { type: integer, example: 3 }
                      newPosition: { type: integer, example: 1 }
      responses:
        '200':
          description: Queue priority successfully updated

  /centers/{id}/availability:
    get:
      summary: Check center operational availability & slot capacity
      description: Returns whether the center is available today, current queue length, available slots, and predicted wait time for a new farmer.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          example: up_c1
        - name: date
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: Center availability status

  /farmer/token:
    post:
      summary: Generate a unique procurement token
      description: Generates a unique token for a farmer, validates center slots, enforces duplicate prevention, enters queue, and triggers ML waiting time calculation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [farmerId, centerId, cropId, cropName, estimatedQuantityQtl]
              properties:
                farmerId: { type: string, example: "FARMER-884120" }
                centerId: { type: string, example: "up_c1" }
                cropId: { type: string, example: "wheat" }
                cropName: { type: string, example: "Wheat" }
                category: { type: string, example: "cereals" }
                estimatedQuantityQtl: { type: number, example: 40.0 }
                vehicleType: { type: string, enum: [Tractor-Trolley, Mini-Truck, Pickup, Bullock-Cart, Truck], example: "Tractor-Trolley" }
                vehicleNumber: { type: string, example: "UP-14-BT-9920" }
                scheduledDate: { type: string, format: date, example: "2026-09-05" }
                scheduledSlotTime: { type: string, example: "10:30 AM" }
      responses:
        '201':
          description: Token generated with queue position & ML predicted waiting time
        '400':
          description: Validation error or center full or duplicate token

  /farmer/status:
    get:
      summary: Get live status of farmer token
      description: Returns token status, current queue position, number of vehicles ahead, procurement stages, and estimated waiting time.
      parameters:
        - name: tokenNumber
          in: query
          schema: { type: string }
          example: "UP-GZB-2026-0003"
        - name: farmerId
          in: query
          schema: { type: string }
          example: "FARMER-884120"
        - name: tokenId
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: Farmer token status and queue metrics

  /farmer/waiting-time:
    get:
      summary: Calculate / predict waiting time
      description: Uses queue position, active bays, crop type, and ML model to calculate waiting minutes.
      parameters:
        - name: tokenNumber
          in: query
          schema: { type: string }
        - name: centerId
          in: query
          schema: { type: string }
          example: "up_c1"
        - name: queuePosition
          in: query
          schema: { type: integer }
          example: 3
        - name: cropId
          in: query
          schema: { type: string }
          example: "wheat"
      responses:
        '200':
          description: Predicted wait time and confidence score

  /procurement/status:
    put:
      summary: Update procurement status (Admin only)
      description: Transition token status (WAITING -> CALLED -> IN_PROCUREMENT -> COMPLETED, or WAITING/CALLED -> CANCELLED). Automatically re-indexes the queue and triggers farmer notifications.
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tokenId, newStatus]
              properties:
                tokenId: { type: string, example: "UP-GZB-2026-0003" }
                newStatus: { type: string, enum: [WAITING, CALLED, IN_PROCUREMENT, COMPLETED, CANCELLED], example: "CALLED" }
                reason: { type: string, example: "Ready at Gate 1" }
                inspectorName: { type: string, example: "Inspector Rajveer" }
      responses:
        '200':
          description: Status updated and queue re-ordered

  /notifications:
    post:
      summary: Dispatch notification to a farmer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [farmerId, centerId, eventType, title, message]
              properties:
                farmerId: { type: string, example: "FARMER-884120" }
                centerId: { type: string, example: "up_c1" }
                eventType: { type: string, example: "TURN_APPROACHING" }
                title: { type: string, example: "Turn Approaching" }
                message: { type: string, example: "Please proceed to Gate 1" }
      responses:
        '201':
          description: Notification dispatched
    get:
      summary: Fetch notification inbox
      parameters:
        - name: farmerId
          in: query
          schema: { type: string }
        - name: centerId
          in: query
          schema: { type: string }
      responses:
        '200':
          description: List of notifications

  /auth/login:
    post:
      summary: Admin / Officer login for JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, example: "admin" }
                password: { type: string, example: "MandiAdmin#2026" }
      responses:
        '200':
          description: JWT token returned

  /events:
    get:
      summary: Real-Time Server-Sent Events (SSE) stream
      description: Subscribe to live events (queue updates, farmer calls, stage transitions).
      parameters:
        - name: centerId
          in: query
          schema: { type: string }
      responses:
        '200':
          description: SSE text/event-stream stream

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key